♻️ Organize some reusable components
