+++ modules/ppcp-settings/resources/js/Components/ReusableComponents/Elements/Action.js
@@ -1,0 +1,5 @@
+const Action = ( { children } ) => (
+	<div className="ppcp-r-settings-block__action">{ children }</div>
+);
+
+export default Action;
