+++ modules/ppcp-settings/resources/js/Components/ReusableComponents/Elements/Content.js
@@ -1,0 +1,16 @@
+import classNames from 'classnames';
+
+const Content = ( { children, className = '', id = '' } ) => {
+	const elementClasses = classNames(
+		'ppcp-r-settings-card__content',
+		className
+	);
+
+	return (
+		<div id={ id } className={ elementClasses }>
+			{ children }
+		</div>
+	);
+};
+
+export default Content;
