+++ modules/ppcp-settings/resources/js/Components/ReusableComponents/Elements/ContentWrapper.js
@@ -1,0 +1,5 @@
+const ContentWrapper = ( { children } ) => (
+	<div className="ppcp-r-settings-card__content-wrapper">{ children }</div>
+);
+
+export default ContentWrapper;
