+++ modules/ppcp-settings/resources/js/Components/ReusableComponents/Elements/SupplementaryLabel.js
@@ -1,0 +1,7 @@
+const SupplementaryLabel = ( { children } ) => (
+	<span className="ppcp-r-settings-block__supplementary-title-label">
+		{ children }
+	</span>
+);
+
+export default SupplementaryLabel;
