+++ modules/ppcp-settings/resources/js/Components/ReusableComponents/Elements/TitleWrapper.js
@@ -1,0 +1,5 @@
+const TitleWrapper = ( { children } ) => (
+	<span className="ppcp-r-settings-block__title-wrapper">{ children }</span>
+);
+
+export default TitleWrapper;
