+++ modules/ppcp-settings/resources/js/Components/ReusableComponents/Elements/index.js
@@ -1,0 +1,8 @@
+export { default as Action } from './Action';
+export { default as Content } from './Content';
+export { default as ContentWrapper } from './ContentWrapper';
+export { default as Description } from './Description';
+export { default as Header } from './Header';
+export { default as SupplementaryLabel } from './SupplementaryLabel';
+export { default as Title } from './Title';
+export { default as TitleWrapper } from './TitleWrapper';
