+++ modules/ppcp-settings/resources/js/Components/ReusableComponents/SettingsBlocks/index.js
@@ -1,4 +1,3 @@
-export { default as SettingsBlock } from './SettingsBlock';
 export { default as ButtonSettingsBlock } from './ButtonSettingsBlock';
 export { default as InputSettingsBlock } from './InputSettingsBlock';
 export { default as SelectSettingsBlock } from './SelectSettingsBlock';
@@ -7,14 +6,3 @@ export { default as ToggleSettingsBlock } from './ToggleSettingsBlock';
 export { default as RadioSettingsBlock } from './RadioSettingsBlock';
 export { default as PaymentMethodsBlock } from './PaymentMethodsBlock';
 export { default as PaymentMethodItemBlock } from './PaymentMethodItemBlock';
-
-export {
-	Title,
-	TitleWrapper,
-	SupplementaryLabel,
-	Description,
-	Action,
-	Content,
-	ContentWrapper,
-	Header,
-} from './SettingsBlockElements';
