+++ modules/ppcp-settings/resources/js/Components/Screens/Overview/TabSettingsElements/Blocks/Troubleshooting/HooksTableBlock.js
@@ -1,9 +1,8 @@
 import { __ } from '@wordpress/i18n';
+
 import { CommonHooks } from '../../../../../../data';
-import {
-	SettingsBlock,
-	Title,
-} from '../../../../../ReusableComponents/SettingsBlocks';
+import SettingsBlock from '../../../../../ReusableComponents/SettingsBlock';
+import { Title } from '../../../../../ReusableComponents/Elements';
 
 const HooksTableBlock = () => {
 	const { webhooks } = CommonHooks.useWebhooks();
