+++ modules/ppcp-settings/resources/js/Components/Screens/Overview/TabSettingsElements/Blocks/Troubleshooting/ResubscribeBlock.js
@@ -1,9 +1,10 @@
-import { useState } from '@wordpress/element';
-import { STORE_NAME } from '../../../../../../data/common';
-import { ButtonSettingsBlock } from '../../../../../ReusableComponents/SettingsBlocks';
 import { __ } from '@wordpress/i18n';
 import { useDispatch } from '@wordpress/data';
+import { useState } from '@wordpress/element';
 import { store as noticesStore } from '@wordpress/notices';
+
+import { STORE_NAME } from '../../../../../../data/common';
+import { ButtonSettingsBlock } from '../../../../../ReusableComponents/SettingsBlocks';
 import {
 	NOTIFICATION_ERROR,
 	NOTIFICATION_SUCCESS,
